requirements.txt
flask
face_recognition
numpy
pillow